Move the synchronous verification process to a dedicated submodule
